+++ components/gozen_growth/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@pipedream/gozen_growth",
-  "version": "0.0.1",
+  "version": "0.1.0",
   "description": "Pipedream GoZen Growth Components",
   "main": "gozen_growth.app.mjs",
   "keywords": [
@@ -11,5 +11,9 @@
   "author": "Pipedream <support@pipedream.com> (https://pipedream.com/)",
   "publishConfig": {
     "access": "public"
+  },
+  "dependencies": {
+    "@pipedream/platform": "^3.0.1"
   }
 }
+
